DP-920 inlined copy-paste tests into HybridEditorTest; made expression acessible in BaseHybridEditorEditingTest
